Clean up Configuration files one class per file.
