+++ src/PresenceLight.Core/Configuration/AvailabilityStatus.cs
@@ -1,0 +1,9 @@
+﻿namespace PresenceLight.Core
+{
+    public class AvailabilityStatus
+    {
+        public bool Disabled { get; set; }
+
+        public string? Colour { get; set; }
+    }
+}
