+++ src/PresenceLight.Core/Configuration/CustomApi.cs
@@ -1,12 +1,4 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
-using System.Text.Json.Serialization;
-
-using Microsoft.Graph.TermStore;
-
-using Newtonsoft.Json;
-
-namespace PresenceLight.Core
+﻿namespace PresenceLight.Core
 {
     public class CustomApi : BaseLight
     {
@@ -66,37 +58,5 @@ namespace PresenceLight.Core
 
 
         public CustomApiSetting CustomApiActivityOff { get; set; }
-    }
-
-    public class LIFX : BaseLight
-    {
-        public string? LIFXApiKey { get; set; }
-
-        public string? LIFXClientId { get; set; }
-
-        public string? LIFXClientSecret { get; set; }
-    }
-
-    public class Hue : BaseLight
-    {
-        public string? RemoteHueClientId { get; set; }
-
-        public string? RemoteHueClientAppName { get; set; }
-
-        public string? RemoteHueClientSecret { get; set; }
-        public string? HueApiKey { get; set; }
-
-        [Required]
-        [RegularExpression(@"\b\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}\b",
-        ErrorMessage = "Not a valid IP Address")]
-        public string? HueIpAddress { get; set; }
-
-        public bool UseRemoteApi { get; set; }
-
-        public string RemoteBridgeId { get; set; }
-    }
-
-    public class Yeelight : BaseLight
-    {
     }
 }
