+++ src/PresenceLight.Core/Configuration/CustomApiSetting.cs
@@ -1,0 +1,8 @@
+﻿namespace PresenceLight.Core
+{
+    public class CustomApiSetting
+    {
+        public string? Method { get; set; }
+        public string? Uri { get; set; }
+    }
+}
