+++ src/PresenceLight.Core/Configuration/LIFX.cs
@@ -1,0 +1,11 @@
+﻿namespace PresenceLight.Core
+{
+    public class LIFX : BaseLight
+    {
+        public string? LIFXApiKey { get; set; }
+
+        public string? LIFXClientId { get; set; }
+
+        public string? LIFXClientSecret { get; set; }
+    }
+}
