+++ src/PresenceLight.Core/Configuration/Yeelight.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Text.Json.Serialization;
+
+using Microsoft.Graph.TermStore;
+
+namespace PresenceLight.Core
+{
+
+    public class Yeelight : BaseLight
+    {
+    }
+}
