+++ src/PresenceLight.Core/PresenceLight.Core.csproj
@@ -13,6 +13,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="HADotNet.Core" Version="1.5.1" />
     <PackageReference Include="LifxCloud" Version="1.1.12" />
     <PackageReference Include="MediatR" Version="9.0.0" />
     <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="9.0.0" />
